Move common functionality into hostcalls mod
