+++ src/sys/unix/hostcalls_impl/misc.rs
@@ -9,71 +9,60 @@ use nix::convert_ioctl_res;
 use nix::libc::{self, c_int};
 use std::cmp;
 use std::time::SystemTime;
-use wasi_common_cbindgen::wasi_common_cbindgen;
 
-#[wasi_common_cbindgen]
-pub fn clock_res_get(
-    memory: &mut [u8],
-    clock_id: wasm32::__wasi_clockid_t,
-    resolution_ptr: wasm32::uintptr_t,
-) -> wasm32::__wasi_errno_t {
+pub(crate) fn clock_res_get(
+    clock_id: host::__wasi_clockid_t,
+) -> Result<host::__wasi_timestamp_t, host::__wasi_errno_t> {
     // convert the supported clocks to the libc types, or return EINVAL
-    let clock_id = match dec_clockid(clock_id) {
+    let clock_id = match clock_id {
         host::__WASI_CLOCK_REALTIME => libc::CLOCK_REALTIME,
         host::__WASI_CLOCK_MONOTONIC => libc::CLOCK_MONOTONIC,
         host::__WASI_CLOCK_PROCESS_CPUTIME_ID => libc::CLOCK_PROCESS_CPUTIME_ID,
         host::__WASI_CLOCK_THREAD_CPUTIME_ID => libc::CLOCK_THREAD_CPUTIME_ID,
-        _ => return wasm32::__WASI_EINVAL,
+        _ => return Err(host::__WASI_EINVAL),
     };
 
     // no `nix` wrapper for clock_getres, so we do it ourselves
     let mut timespec = unsafe { std::mem::uninitialized::<libc::timespec>() };
     let res = unsafe { libc::clock_getres(clock_id, &mut timespec as *mut libc::timespec) };
     if res != 0 {
-        return host_impl::errno_from_nix(nix::errno::Errno::last());
+        return Err(host_impl::errno_from_nix(nix::errno::Errno::last()));
     }
 
-    // convert to nanoseconds, returning EOVERFLOW in case of overflow; this is freelancing a bit
-    // from the spec but seems like it'll be an unusual situation to hit
+    // convert to nanoseconds, returning EOVERFLOW in case of overflow;
+    // this is freelancing a bit from the spec but seems like it'll
+    // be an unusual situation to hit
     (timespec.tv_sec as host::__wasi_timestamp_t)
         .checked_mul(1_000_000_000)
         .and_then(|sec_ns| sec_ns.checked_add(timespec.tv_nsec as host::__wasi_timestamp_t))
-        .map_or(wasm32::__WASI_EOVERFLOW, |resolution| {
+        .map_or(Err(host::__WASI_EOVERFLOW), |resolution| {
             // a supported clock can never return zero; this case will probably never get hit, but
             // make sure we follow the spec
             if resolution == 0 {
-                wasm32::__WASI_EINVAL
+                Err(host::__WASI_EINVAL)
             } else {
-                enc_timestamp_byref(memory, resolution_ptr, resolution)
-                    .map(|_| wasm32::__WASI_ESUCCESS)
-                    .unwrap_or_else(|e| e)
+                Ok(resolution)
             }
         })
 }
 
-#[wasi_common_cbindgen]
-pub fn clock_time_get(
-    memory: &mut [u8],
-    clock_id: wasm32::__wasi_clockid_t,
-    // ignored for now, but will be useful once we put optional limits on precision to reduce side
-    // channels
-    _precision: wasm32::__wasi_timestamp_t,
-    time_ptr: wasm32::uintptr_t,
-) -> wasm32::__wasi_errno_t {
+pub(crate) fn clock_time_get(
+    clock_id: host::__wasi_clockid_t,
+) -> Result<host::__wasi_timestamp_t, host::__wasi_errno_t> {
     // convert the supported clocks to the libc types, or return EINVAL
-    let clock_id = match dec_clockid(clock_id) {
+    let clock_id = match clock_id {
         host::__WASI_CLOCK_REALTIME => libc::CLOCK_REALTIME,
         host::__WASI_CLOCK_MONOTONIC => libc::CLOCK_MONOTONIC,
         host::__WASI_CLOCK_PROCESS_CPUTIME_ID => libc::CLOCK_PROCESS_CPUTIME_ID,
         host::__WASI_CLOCK_THREAD_CPUTIME_ID => libc::CLOCK_THREAD_CPUTIME_ID,
-        _ => return wasm32::__WASI_EINVAL,
+        _ => return Err(host::__WASI_EINVAL),
     };
 
     // no `nix` wrapper for clock_getres, so we do it ourselves
     let mut timespec = unsafe { std::mem::uninitialized::<libc::timespec>() };
     let res = unsafe { libc::clock_gettime(clock_id, &mut timespec as *mut libc::timespec) };
     if res != 0 {
-        return host_impl::errno_from_nix(nix::errno::Errno::last());
+        return Err(host_impl::errno_from_nix(nix::errno::Errno::last()));
     }
 
     // convert to nanoseconds, returning EOVERFLOW in case of overflow; this is freelancing a bit
@@ -81,33 +70,13 @@ pub fn clock_time_get(
     (timespec.tv_sec as host::__wasi_timestamp_t)
         .checked_mul(1_000_000_000)
         .and_then(|sec_ns| sec_ns.checked_add(timespec.tv_nsec as host::__wasi_timestamp_t))
-        .map_or(wasm32::__WASI_EOVERFLOW, |time| {
-            enc_timestamp_byref(memory, time_ptr, time)
-                .map(|_| wasm32::__WASI_ESUCCESS)
-                .unwrap_or_else(|e| e)
-        })
+        .map_or(Err(host::__WASI_EOVERFLOW), |time| Ok(time))
 }
 
-#[wasi_common_cbindgen]
-pub fn poll_oneoff(
-    memory: &mut [u8],
-    input: wasm32::uintptr_t,
-    output: wasm32::uintptr_t,
-    nsubscriptions: wasm32::size_t,
-    nevents: wasm32::uintptr_t,
-) -> wasm32::__wasi_errno_t {
-    if nsubscriptions as u64 > wasm32::__wasi_filesize_t::max_value() {
-        return wasm32::__WASI_EINVAL;
-    }
-    enc_pointee(memory, nevents, 0).unwrap();
-    let input_slice =
-        dec_slice_of::<wasm32::__wasi_subscription_t>(memory, input, nsubscriptions).unwrap();
-
-    let input: Vec<_> = input_slice.iter().map(|x| dec_subscription(x)).collect();
-
-    let output_slice =
-        dec_slice_of_mut::<wasm32::__wasi_event_t>(memory, output, nsubscriptions).unwrap();
-
+pub(crate) fn poll_oneoff(
+    input: Vec<Result<host::__wasi_subscription_t, host::__wasi_errno_t>>,
+    output_slice: &mut [wasm32::__wasi_event_t],
+) -> Result<wasm32::size_t, host::__wasi_errno_t> {
     let timeout = input
         .iter()
         .filter_map(|event| match event {
@@ -135,7 +104,7 @@ pub fn poll_oneoff(
         })
         .collect();
     if fd_events.is_empty() && timeout.is_none() {
-        return wasm32::__WASI_ESUCCESS;
+        return Ok(0);
     }
     let mut poll_fds: Vec<_> = fd_events
         .iter()
@@ -163,7 +132,7 @@ pub fn poll_oneoff(
                 if nix::errno::Errno::last() == nix::errno::Errno::EINTR {
                     continue;
                 }
-                return host_impl::errno_from_nix(nix::errno::Errno::last());
+                return Err(host_impl::errno_from_nix(nix::errno::Errno::last()));
             }
             Ok(ready) => break ready as usize,
         }
@@ -174,16 +143,13 @@ pub fn poll_oneoff(
         let events = fd_events.iter().zip(poll_fds.iter()).take(ready);
         poll_oneoff_handle_fd_event(output_slice, events)
     };
-    if let Err(e) = enc_pointee(memory, nevents, events_count) {
-        return enc_errno(e);
-    }
-    wasm32::__WASI_ESUCCESS
+
+    Ok(events_count)
 }
 
-#[wasi_common_cbindgen]
-pub fn sched_yield() -> wasm32::__wasi_errno_t {
+pub(crate) fn sched_yield() -> Result<(), host::__wasi_errno_t> {
     unsafe { libc::sched_yield() };
-    wasm32::__WASI_ESUCCESS
+    Ok(())
 }
 
 // define the `fionread()` function, equivalent to `ioctl(fd, FIONREAD, *bytes)`
