+++ src/sys/unix/hostcalls_impl/mod.rs
@@ -3,11 +3,9 @@
 mod fs;
 mod fs_helpers;
 mod misc;
-mod sock;
 
 use super::fdentry;
 use super::host_impl;
 
-pub use self::fs::*;
-pub use self::misc::*;
-pub use self::sock::*;
+pub(crate) use self::fs::*;
+pub(crate) use self::misc::*;
