+++ src/sys/windows/hostcalls_impl/mod.rs
@@ -2,10 +2,9 @@
 //! [WASI](https://github.com/CraneStation/wasmtime-wasi/blob/wasi/docs/WASI-overview.md).
 mod fs;
 mod misc;
-mod sock;
 
+use super::fdentry;
 use super::host_impl;
 
-pub use self::fs::*;
-pub use self::misc::*;
-pub use self::sock::*;
+pub(crate) use self::fs::*;
+pub(crate) use self::misc::*;
